Fix sql patches for case when rottenlinks.rl_externallink is a primary key (#30)
In some cases rottenlinks can have rl_externallink as a primary key, but in
some cases it is not. In the first case the database updater throws
SQL exception (it can't create primary key if it already exists)
the removePrimaryKey stored procedure removes primary key in the
rottenlinks table if it exists.